backend/requirements.txt
attrs==21.2.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==8.1.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
Flask==2.2.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.1
more-itertools==8.10.0
mysql-connector-python==8.0.31
netifaces==0.11.0
oauthlib==3.2.0
protobuf==3.20.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.1
python-dotenv==0.19.2
PyYAML==5.4.1
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
texttable==1.6.4
ubuntu-advantage-tools==27.12
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
websocket-client==1.2.3
Werkzeug==2.2.2
zipp==1.0.0
